+++ src/Service/OpenExchangeRates.php
@@ -131,7 +131,7 @@ class OpenExchangeRates extends HistoricalService
         if ($data['base'] === $currencyPair->getBaseCurrency()
             && isset($data['rates'][$currencyPair->getQuoteCurrency()])
         ) {
-            return new ExchangeRate($data['rates'][$currencyPair->getQuoteCurrency()], __CLASS__, $date);
+            return new ExchangeRate((float) ($data['rates'][$currencyPair->getQuoteCurrency()]), __CLASS__, $date);
         }
 
         throw new UnsupportedCurrencyPairException($currencyPair, $this);
